move out ByteArray from TelegramBuffer to a separate file
